backend/requirements.txt
fastapi==0.115.11
uvicorn==0.32.1
qdrant-client==1.16.1
psycopg2-binary==2.9.11
sentence-transformers==3.4.1
pydantic==2.11.7
python-dotenv==1.0.1
requests==2.32.3
numpy>=1.21.0
openai==1.54.4
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1